+++ modules/client/CHANGELOG.md
@@ -21,6 +21,7 @@ TEMPLATE:
 - Added `initializeFrom` encoders and decoders 
 - Support for ERC721 deposits and withdrawals
 - Added `getProtocolVersion` function
+- Support for baseGoerli network with the new client-common version
 ### Fixes
 - Fix status calculation for token voting proposals
 - Make the `network` parameter required on `getPluginInstallItem`